api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
motor>=3.3.0
pymongo>=4.0.0
python-dotenv>=1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]>=3.3.0
bcrypt==3.2.2
python-multipart>=0.0.6
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0 